Remove reference dataflow analysis in goggles
It doens't fit well into python since python heavily relies on
mutation.

Ideally, it would operate like a DAG of cells, with snapshots of
variables taken before they were mutated so we could track the deltas
and compare them, but I don't know of an easy way of intercepting
mutation in the python VM. Maybe monkey patch everything or intercept at
the bytecode level?

In any case, making each cell operate as an independent experiment is
useful and actually is a slightly simpler model to reason about I think.
I'll save dataflow for a rainy day in the future...